working window and image - imported olivec for drawing

+++ shm.c
@@ -4,7 +4,7 @@
 #include <time.h>
 #include <unistd.h>
 
-int create_shm_file(size_t size) {
+int create_shm_file(int size) {
     int fd;
 	int retries = 100;
 	do {
